+++ pkg/engine/context/context.go
@@ -6,9 +6,10 @@ import (
 	"sync"
 
 	jsonpatch "github.com/evanphx/json-patch/v5"
-	urkyverno "github.com/kyverno/kyverno/api/kyverno/v1beta1"
+	kyvernov1 "github.com/kyverno/kyverno/api/kyverno/v1"
+	kyvernov1beta1 "github.com/kyverno/kyverno/api/kyverno/v1beta1"
 	pkgcommon "github.com/kyverno/kyverno/pkg/common"
-	kubeutils "github.com/kyverno/kyverno/pkg/utils/kube"
+	apiutils "github.com/kyverno/kyverno/pkg/utils/api"
 	"github.com/pkg/errors"
 	admissionv1 "k8s.io/api/admission/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -50,7 +51,7 @@ type Interface interface {
 	AddOldResource(data map[string]interface{}) error
 
 	// AddUserInfo merges userInfo json under kyverno.userInfo
-	AddUserInfo(userInfo urkyverno.RequestInfo) error
+	AddUserInfo(userInfo kyvernov1beta1.RequestInfo) error
 
 	// AddServiceAccount merges ServiceAccount types
 	AddServiceAccount(userName string) error
@@ -62,17 +63,17 @@ type Interface interface {
 	AddElement(data interface{}, index int) error
 
 	// AddImageInfo adds image info to the context
-	AddImageInfo(info kubeutils.ImageInfo) error
+	AddImageInfo(info apiutils.ImageInfo) error
 
 	// AddImageInfos adds image infos to the context
 	AddImageInfos(resource *unstructured.Unstructured) error
 
 	// ImageInfo returns image infos present in the context
-	ImageInfo() map[string]map[string]kubeutils.ImageInfo
+	ImageInfo() map[string]map[string]apiutils.ImageInfo
 
 	// GenerateCustomImageInfo returns image infos as defined by a custom image extraction config
 	// and updates the context
-	GenerateCustomImageInfo(resource *unstructured.Unstructured, imageExtractorConfigs kubeutils.ImageExtractorConfigs) (map[string]map[string]kubeutils.ImageInfo, error)
+	GenerateCustomImageInfo(resource *unstructured.Unstructured, imageExtractorConfigs kyvernov1.ImageExtractorConfigs) (map[string]map[string]apiutils.ImageInfo, error)
 
 	// Checkpoint creates a copy of the current internal state and pushes it into a stack of stored states.
 	Checkpoint()
@@ -94,7 +95,7 @@ type context struct {
 	mutex              sync.RWMutex
 	jsonRaw            []byte
 	jsonRawCheckpoints [][]byte
-	images             map[string]map[string]kubeutils.ImageInfo
+	images             map[string]map[string]apiutils.ImageInfo
 }
 
 // NewContext returns a new context
@@ -166,7 +167,7 @@ func (ctx *context) AddOldResource(data map[string]interface{}) error {
 }
 
 // AddUserInfo adds userInfo at path request.userInfo
-func (ctx *context) AddUserInfo(userRequestInfo urkyverno.RequestInfo) error {
+func (ctx *context) AddUserInfo(userRequestInfo kyvernov1beta1.RequestInfo) error {
 	return addToContext(ctx, userRequestInfo, "request")
 }
 
@@ -231,7 +232,7 @@ func (ctx *context) AddElement(data interface{}, index int) error {
 	return addToContext(ctx, data)
 }
 
-func (ctx *context) AddImageInfo(info kubeutils.ImageInfo) error {
+func (ctx *context) AddImageInfo(info apiutils.ImageInfo) error {
 	data := map[string]interface{}{
 		"image":    info.String(),
 		"registry": info.Registry,
@@ -247,7 +248,7 @@ func (ctx *context) AddImageInfos(resource *unstructured.Unstructured) error {
 
 	log.Log.V(4).Info("extracting image info", "obj", resource.UnstructuredContent())
 
-	images, err := kubeutils.ExtractImagesFromResource(*resource, nil)
+	images, err := apiutils.ExtractImagesFromResource(*resource, nil)
 	if err != nil {
 		return err
 	}
@@ -261,8 +262,8 @@ func (ctx *context) AddImageInfos(resource *unstructured.Unstructured) error {
 	return addToContext(ctx, images, "images")
 }
 
-func (ctx *context) GenerateCustomImageInfo(resource *unstructured.Unstructured, imageExtractorConfigs kubeutils.ImageExtractorConfigs) (map[string]map[string]kubeutils.ImageInfo, error) {
-	images, err := kubeutils.ExtractImagesFromResource(*resource, imageExtractorConfigs)
+func (ctx *context) GenerateCustomImageInfo(resource *unstructured.Unstructured, imageExtractorConfigs kyvernov1.ImageExtractorConfigs) (map[string]map[string]apiutils.ImageInfo, error) {
+	images, err := apiutils.ExtractImagesFromResource(*resource, imageExtractorConfigs)
 	if err != nil {
 		return nil, errors.Wrapf(err, "failed to extract images")
 	}
@@ -275,7 +276,7 @@ func (ctx *context) GenerateCustomImageInfo(resource *unstructured.Unstructured,
 	return images, addToContext(ctx, images, "images")
 }
 
-func (ctx *context) ImageInfo() map[string]map[string]kubeutils.ImageInfo {
+func (ctx *context) ImageInfo() map[string]map[string]apiutils.ImageInfo {
 	return ctx.images
 }