+++ pkg/policycache/informer.go
@@ -5,10 +5,10 @@ import (
 	"reflect"
 	"sync/atomic"
 
-	"github.com/go-logr/logr"
 	kyverno "github.com/kyverno/kyverno/api/kyverno/v1"
 	kyvernoinformer "github.com/kyverno/kyverno/pkg/client/informers/externalversions/kyverno/v1"
 	kyvernolister "github.com/kyverno/kyverno/pkg/client/listers/kyverno/v1"
+	kubeutils "github.com/kyverno/kyverno/pkg/utils/kube"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/client-go/tools/cache"
@@ -21,22 +21,15 @@ import (
 // policies based on types (Mutate/ValidateEnforce/Generate/imageVerify).
 type Controller struct {
 	Cache      Interface
-	log        logr.Logger
 	cpolLister kyvernolister.ClusterPolicyLister
 	polLister  kyvernolister.PolicyLister
 	pCounter   int64
 }
 
 // NewPolicyCacheController create a new PolicyController
-func NewPolicyCacheController(
-	pInformer kyvernoinformer.ClusterPolicyInformer,
-	nspInformer kyvernoinformer.PolicyInformer,
-	log logr.Logger,
-) *Controller {
-
+func NewPolicyCacheController(pInformer kyvernoinformer.ClusterPolicyInformer, nspInformer kyvernoinformer.PolicyInformer) *Controller {
 	pc := Controller{
-		Cache: newPolicyCache(log, pInformer.Lister(), nspInformer.Lister()),
-		log:   log,
+		Cache: newPolicyCache(pInformer.Lister(), nspInformer.Lister()),
 	}
 
 	// ClusterPolicy Informer
@@ -75,8 +68,12 @@ func (c *Controller) updatePolicy(old, cur interface{}) {
 }
 
 func (c *Controller) deletePolicy(obj interface{}) {
-	p := obj.(*kyverno.ClusterPolicy)
-	c.Cache.remove(p)
+	p, ok := kubeutils.GetObjectWithTombstone(obj).(*kyverno.ClusterPolicy)
+	if ok {
+		c.Cache.remove(p)
+	} else {
+		logger.Info("Failed to get deleted object, the deleted policy cannot be removed from the cache", "obj", obj)
+	}
 }
 
 // addNsPolicy - Add Policy to cache
@@ -97,13 +94,16 @@ func (c *Controller) updateNsPolicy(old, cur interface{}) {
 
 // deleteNsPolicy - Delete Policy from cache
 func (c *Controller) deleteNsPolicy(obj interface{}) {
-	p := obj.(*kyverno.Policy)
-	c.Cache.remove(p)
+	p, ok := kubeutils.GetObjectWithTombstone(obj).(*kyverno.Policy)
+	if ok {
+		c.Cache.remove(p)
+	} else {
+		logger.Info("Failed to get deleted object, the deleted cluster policy cannot be removed from the cache", "obj", obj)
+	}
 }
 
 // CheckPolicySync wait until the internal policy cache is fully loaded
 func (c *Controller) CheckPolicySync(stopCh <-chan struct{}) {
-	logger := c.log
 	logger.Info("starting")
 
 	policies := []kyverno.PolicyInterface{}