+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `status.ready` of the policy is deprecated in favor of `policy.IsReady()`. The implementation was changed to use `status.conditions` that offer more flexibility. The `status.ready` will be kept for a couple of releases until we remove it in the future.
 - Deprecated flags have been removed.
+- Flags that were overlapping with config map based configuration were removed (`filterK8sResources`, `excludeGroupRole`, `excludeUsername`). They can now be configured using the config map only.
 
 ## v1.6.0-rc1
 ### Note